rename modelzoo/apis to modelzoo/api. Make it possible to choose video adaptation in the inference api
